requirements.txt
Django==3.2.13
psycopg2-binary==2.9.3
boto3==1.24.10
django-rq
django-allauth
django-oauth-toolkit
pkce
djangorestframework
PyGithub
